Reorganise suggestion package
