+++ src/test/java/com/notably/logic/suggestion/commands/HelpSuggestionCommandTest.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
+import com.notably.logic.commands.suggestion.HelpSuggestionCommand;
 import com.notably.model.Model;
 import com.notably.model.ModelManager;
 import com.notably.model.block.BlockModel;
