+++ src/test/java/com/notably/logic/suggestion/commands/OpenSuggestionCommandTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.Test;
 
 import com.notably.commons.path.AbsolutePath;
 import com.notably.commons.path.exceptions.InvalidPathException;
+import com.notably.logic.commands.suggestion.OpenSuggestionCommand;
 import com.notably.model.Model;
 import com.notably.model.ModelManager;
 import com.notably.model.block.Block;
